requirements.txt
fastapi
uvicorn
opencv-python
pillow
torch
git+https://github.com/openai/CLIP.git
faiss-cpu
numpy